requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
Django==5.0.4
django-filter==24.2
django-jazzmin==3.0.1
django-suit==0.2.28
djangorestframework==3.15.1
google-api-core==2.24.2
google-api-python-client==2.171.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
Markdown==3.6
oauthlib==3.2.2
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.4.0
